--- uniflow/extract/__init__.py
@@ -1,5 +1,0 @@
-"""Extract __init__ Module."""
-from uniflow.extract.flow.extract_pdf_flow import ExtractPDFFlow  # noqa: F401;
-from uniflow.extract.flow.extract_txt_flow import ExtractTxtFlow  # noqa: F401, F403
-
-__all__ = ["ExtractTxtFlow", "ExtractPDFFlow"]
