+++ uniflow/flow/extract/load/extract_txt_flow.py
@@ -3,7 +3,7 @@
 from typing import Sequence
 
 from uniflow.constants import EXTRACT
-from uniflow.flow import Flow
+from uniflow.flow.flow import Flow
 from uniflow.node.node import Node
 from uniflow.op.extract.txt_op import ExtractTxtOp, ProcessTxtOp
