+++ uniflow/flow/transform/transform_lmqg_flow.py
@@ -1,11 +1,11 @@
 from typing import Any, Dict, Sequence
 
 from uniflow.constants import TRANSFORM
-from uniflow.flow import Flow
-from uniflow.model.model import Model
+from uniflow.flow.flow import Flow
+from uniflow.flow.transform.model import Model
 from uniflow.node.node import Node
-from uniflow.op.model.model_op import ModelOp
-from uniflow.schema import GuidedPrompt
+from uniflow.op.transform.model_op import ModelOp
+from uniflow.op.prompt_schema import GuidedPrompt
 
 
 class TransformLMQGFlow(Flow):
