+++ uniflow/flow/transform/transform_openai_flow.py
@@ -1,11 +1,12 @@
 """Model Flow Module."""
 from typing import Any, Dict, Sequence
 
-from uniflow.flow import Flow
-from uniflow.model.model import JsonModel, Model
+from uniflow.constants import TRANSFORM
+from uniflow.flow.flow import Flow
+from uniflow.flow.transform.model import JsonModel, Model
 from uniflow.node.node import Node
-from uniflow.op.model.model_op import ModelOp
-from uniflow.schema import GuidedPrompt
+from uniflow.op.transform.model_op import ModelOp
+from uniflow.op.prompt_schema import GuidedPrompt
 
 
 class OpenAIModelFlow(Flow):
@@ -48,3 +49,7 @@ class OpenAIModelFlow(Flow):
             Sequence[Node]: Nodes after running.
         """
         return self._model_op(nodes)
+
+
+class TransformOpenAIFlow(OpenAIModelFlow):
+    TAG = TRANSFORM
