+++ uniflow/op/transform/model_op.py
@@ -2,7 +2,7 @@
 import copy
 from typing import Sequence
 
-from uniflow.model.model import Model
+from uniflow.flow.transform.model import Model
 from uniflow.node.node import Node
 from uniflow.op.op import Op
