+++ uniflow/op/transform/server.py
@@ -8,9 +8,9 @@ from typing import Any, Dict, List, Mapping, Tuple
 from tqdm.auto import tqdm
 
 from uniflow.constants import TRANSFORM
-from uniflow.flow_factory import FlowFactory
+from uniflow.flow.flow_factory import FlowFactory
 from uniflow.op.op import OpScope
-from uniflow.transform.config import TransformConfig
+from uniflow.op.config import TransformConfig
 
 
 class Server:
