--- uniflow/transform/flow/transform_huggingface_flow.py
@@ -1,6 +1,0 @@
-from uniflow.constants import TRANSFORM
-from uniflow.model.flow.model_huggingface_flow import HuggingFaceModelFlow
-
-
-class TransformHuggingFaceFlow(HuggingFaceModelFlow):
-    TAG = TRANSFORM
