--- uniflow/transform/flow/transform_openai_flow.py
@@ -1,6 +1,0 @@
-from uniflow.constants import TRANSFORM
-from uniflow.model.flow.model_openai_flow import OpenAIModelFlow
-
-
-class TransformOpenAIFlow(OpenAIModelFlow):
-    TAG = TRANSFORM
